factor out viqe client http calls.
